requirements.txt
streamlit==1.42.0
anthropic>=0.40.0
pdfplumber==0.11.5
PyMuPDF==1.25.3
reportlab==4.2.5
duckduckgo-search==7.5.0
python-dotenv==1.0.1
cryptography==44.0.1
pytest==8.3.4
